[test] Test all fields from granules are in the DB
This test is designed to catch cases where a granule field is being
parsed from the original file, but it is omitted in COLUMN_TO_FIELD or
in the DB schema and so is not included in the database.
NOTE not all granule fields are read from the granule files, because
there is so much irrelevant/duplicated data there. This test does not
check that all product data dictionary fields are in the DB, because
they are not supposed to be there.
However, the granule file parsers should be the source of truth. If a
field is parsed out of the granule file, it should be deduplicated with
other products and included in the database. This test is designed to
catch cases where that is not happening.

Currently, some fields that should be present in the DB are not.
(Because this test didn't exist!)
It requires a complicated migration to add them, so for now the test
warns about these specific columns but does not fail on them.
If forking this repo and creating a new database from scratch,
these instances should be fixed in the code and the tests updated.