[fix/feat] skip H5 download if parquet file exists
1. Skip granule download if the filtered parquet file for the granule
already exists (even if the granule is not present in the database).
This helps with the estimation of remaining granules/download size for
partially finished regions AND allows us to immediately delete an H5
file once we write the filtered/joined parquet. It also removes
processed granules with no valid data from the granule download size
estimate.
TODO: Insert granules with no valid data into the granules table in order
to record that they have been processed.
2. Delete H5 files immediately once a granule parquet file is processed.
This makes the peak storage footprint more similar to the estimate
provided to the user.
3. Restrict to pre-storage (with warning)
4. Fix bug where empty placeholder parquet files were invalid (no
geometry column)